会社選択のコンポーネントを分離
